// api/models/SetMetrics.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SetMetrics = {
    total: number;
    wins: number;
    losses: number;
    bozos: number;
    pushes: number;
    voids: number;
    curr_win_streak: number;
    curr_loss_streak: number;
    curr_bozo_streak: number;
    longest_win_streak: number;
    longest_loss_streak: number;
    longest_bozo_streak: number;
    win_rate: (number | null);
    bozo_rate: (number | null);
};